# backend/docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact:
    email: support@financetracker.local
    name: API Support
  description: Finance management REST API with Auth0 authentication
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: FinanceTracker API
  version: "1.0"
paths:
...
  /auth/callback:
    get:
      description: Auth0からの認証コールバックを処理し、ユーザー情報を保存します
      parameters:
      - description: 認証コード
        in: query
        name: code
        required: true
        type: string
      - description: CSRF防止用のstate
        in: query
        name: state
        required: true
        type: string
      responses:
        "200":
          description: ユーザー情報とトークン
          schema:
            additionalProperties: true
            type: object
        "400":
          description: リクエストエラー
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 認証エラー
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部エラー
          schema:
            additionalProperties: true
            type: object
      summary: Auth0コールバック
      tags:
      - auth
  /auth/check:
    get:
      description: HttpOnlyクッキーからトークンを読み取り、認証状態を確認します
      responses:
        "200":
          description: 認証成功
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 認証失敗
          schema:
            additionalProperties: true
            type: object
      summary: 認証状態チェック
      tags:
      - auth
...
  /auth/token:
    delete:
      description: HttpOnlyクッキーからアクセストークンを削除します
      responses:
        "200":
          description: トークン削除成功
          schema:
            additionalProperties: true
            type: object
      summary: トークン削除
      tags:
      - auth
    post:
      description: Auth0トークンをHttpOnlyクッキーに保存します
      parameters:
      - description: トークンリクエスト
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      responses:
        "200":
          description: トークン設定成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: リクエストエラー
          schema:
            additionalProperties: true
            type: object
      summary: トークン設定
      tags:
      - auth
  /auth/user:
    get:
      description: 認証済みユーザーの情報を取得します
      responses:
        "200":
          description: ユーザー情報
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 認証エラー
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部エラー
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 現在のユーザー情報取得
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"